refactor: rename core auth to oauth
