+++ src/core/oauth/mapGithubProfile.ts
